Don't replace a file if the new file is identical to the existing one in the DAM

+++ packages/api/cms-api/src/dam/files/files.service.ts
@@ -248,31 +248,38 @@ export class FilesService {
                 );
             }
 
-            const { exifData, contentHash, image } = await this.getFileMetadataForUpload(uploadedFile);
-            await this.blobStorageBackendService.upload(uploadedFile, contentHash, this.config.filesDirectory);
+            const uploadedFileMetadata = await this.getFileMetadataForUpload(uploadedFile);
+            const oldAndNewFileAreIdentical = fileToReplace.contentHash === uploadedFileMetadata.contentHash;
 
-            // Check if the current file is the only one using the contentHash before deleting from blob storage
-            if (
-                (await withFilesSelect(this.filesRepository.createQueryBuilder("file"), { contentHash: fileToReplace.contentHash }).getResult())
-                    .length === 1
-            ) {
-                await this.blobStorageBackendService.removeFile(this.config.filesDirectory, createHashedPath(fileToReplace.contentHash));
+            if (!oldAndNewFileAreIdentical) {
+                await this.blobStorageBackendService.upload(uploadedFile, uploadedFileMetadata.contentHash, this.config.filesDirectory);
+
+                // Check if the current file is the only one using the contentHash before deleting from blob storage
+                if (
+                    (await withFilesSelect(this.filesRepository.createQueryBuilder("file"), { contentHash: fileToReplace.contentHash }).getResult())
+                        .length === 1
+                ) {
+                    await this.blobStorageBackendService.removeFile(this.config.filesDirectory, createHashedPath(fileToReplace.contentHash));
+                }
+
+                if (uploadedFileMetadata.image && uploadedFileMetadata.image.width && uploadedFileMetadata.image.height && fileToReplace.image) {
+                    fileToReplace.image.width = uploadedFileMetadata.image.width;
+                    fileToReplace.image.height = uploadedFileMetadata.image.height;
+                    fileToReplace.image.exif = uploadedFileMetadata.exifData;
+                }
+
+                Object.assign(fileToReplace, {
+                    size: uploadedFile.size,
+                    mimetype: uploadedFile.mimetype,
+                    contentHash: uploadedFileMetadata.contentHash,
+                    ...assignData,
+                });
+
+                result = await this.save(fileToReplace);
+            } else {
+                // if old and new files are identical, skip replacement and return the existing file
+                result = fileToReplace;
             }
-
-            if (image && image.width && image.height && fileToReplace.image) {
-                fileToReplace.image.width = image.width;
-                fileToReplace.image.height = image.height;
-                fileToReplace.image.exif = exifData;
-            }
-
-            Object.assign(fileToReplace, {
-                size: uploadedFile.size,
-                mimetype: uploadedFile.mimetype,
-                contentHash,
-                ...assignData,
-            });
-
-            result = await this.save(fileToReplace);
 
             rimraf.sync(uploadedFile.path);
         } catch (e) {
